refactor: finish the rename where the vocabulary still said benchmark
The shared timing package is micro_benchmarking rather than micro, which said nothing outside the package it used to sit in. The evaluation classes, their module and the report they print follow the same distinction the topic rests on: benchmarking measures a machine, evaluation measures this library. Docs capture recaptured for the changed wording.

The lazy-name construction gets the docstring it lacked, stating why the name cannot live in __all__: a wildcard import would getattr it and fire the guard.
