+++ src/counted_float/_core/evaluation/__init__.py
@@ -1,4 +1,4 @@
 """Evaluating the library itself, as opposed to benchmarking the machine it runs on."""
 
-from ._counted_float_benchmark import BenchmarkCountedFloat, BenchmarkFloat, CountedFloatBenchmarkResults
+from ._overhead_evaluation import CountedFloatEvaluation, CountingOverheadResults, FloatEvaluation
 from ._runner import evaluate_counting_overhead
