+++ src/counted_float/_core/evaluation/_overhead_evaluation.py
@@ -1,5 +1,5 @@
 from counted_float._core.counting import CountedFloat
-from counted_float._core.micro import MicroBenchmark
+from counted_float._core.micro_benchmarking import MicroBenchmark
 from counted_float._core.models import JsonReprModel
 from counted_float._core.utils import format_time_duration
 
@@ -7,12 +7,12 @@ from counted_float._core.utils import format_time_duration
 # =================================================================================================
 #  Result class
 # =================================================================================================
-class CountedFloatBenchmarkResults(JsonReprModel):
+class CountingOverheadResults(JsonReprModel):
     float_time_nsec: float
     counted_float_time_nsec: float
 
     def show(self) -> None:
-        print("CountedFloat Benchmark Results:")
+        print("Counting overhead:")
         print(f"  Bisection using float        : {format_time_duration(self.float_time_nsec)} / execution")
         print(f"  Bisection using CountedFloat : {format_time_duration(self.counted_float_time_nsec)} / execution")
         ratio = self.counted_float_time_nsec / self.float_time_nsec
@@ -21,14 +21,14 @@ class CountedFloatBenchmarkResults(JsonReprModel):
 
 
 # =================================================================================================
-#  MicroBenchmarks
+#  The two workloads being compared
 # =================================================================================================
 def _zero_function(x: float) -> float:
     # function for which we want to find a root
     return (x * x * x) - 7.0
 
 
-class BenchmarkFloat(MicroBenchmark):
+class FloatEvaluation(MicroBenchmark):
     def __init__(self) -> None:
         super().__init__(name="float")
         self._n_executions = 1
@@ -57,7 +57,7 @@ class BenchmarkFloat(MicroBenchmark):
                     fb = fmid  # noqa: F841
 
 
-class BenchmarkCountedFloat(MicroBenchmark):
+class CountedFloatEvaluation(MicroBenchmark):
     def __init__(self) -> None:
         super().__init__(name="CountedFloat")
         self._n_executions = 1
@@ -68,7 +68,7 @@ class BenchmarkCountedFloat(MicroBenchmark):
     def _run_benchmark(self) -> None:
         for _ in range(self._n_executions):
             # Execute bisection to find root of _zero_function in interval [-1e50,1e50],
-            # with identical implementation as BenchmarkFloat, except that we initialize a,b as CountedFloats,
+            # with identical implementation as FloatEvaluation, except that we initialize a,b as CountedFloats,
             # which will make sure all the remaining operations are also executed using CountedFloat arithmetic
             a = CountedFloat(-1e50)
             b = CountedFloat(1e50)
