+++ src/counted_float/evaluation/__init__.py
@@ -1,8 +1,8 @@
 """Public evaluation API: measure what counting costs, on this machine, for this workload."""
 
-from counted_float._core.evaluation import CountedFloatBenchmarkResults, evaluate_counting_overhead
+from counted_float._core.evaluation import CountingOverheadResults, evaluate_counting_overhead
 
 __all__ = [
-    "CountedFloatBenchmarkResults",
+    "CountingOverheadResults",
     "evaluate_counting_overhead",
 ]
