+++ tests/micro_benchmarking/test_interleaved_runner.py
@@ -2,8 +2,8 @@ import random
 
 import pytest
 
-from counted_float._core.micro import InterleavedBenchmarkRunner, MicroBenchmark, SliceController
-from counted_float._core.micro._output import output_quiet
+from counted_float._core.micro_benchmarking import InterleavedBenchmarkRunner, MicroBenchmark, SliceController
+from counted_float._core.micro_benchmarking._output import output_quiet
 from counted_float._core.models import MicroBenchmarkResult, SingleRunResult
