+++ tests/micro_benchmarking/test_micro_benchmark.py
@@ -12,8 +12,8 @@ import time
 
 import pytest
 
-from counted_float._core.micro import MicroBenchmark
-from counted_float._core.micro._output import output_quiet
+from counted_float._core.micro_benchmarking import MicroBenchmark
+from counted_float._core.micro_benchmarking._output import output_quiet
 from counted_float._core.models import MicroBenchmarkResult
